Move runVlogGC to x and use it in zero as well (#5468)
